Termination of bugs
