+++ src/python/mini_face/__init__.py
@@ -1,2 +1,5 @@
-from . import gaze, action_unit
-from .common import PredictionMode
+from .api import __doc__, __version__
+from .api import *
+from .gaze import GazeFeatureExtractor
+from .action_unit import AUFeatureExtractor
+from .mode import PredictionMode
