--- src/python/mini_face/action_unit/__init__.py
@@ -1,1 +1,0 @@
-from ._action_unit import Extractor, Result
