--- src/python/mini_face/common/__init__.py
@@ -1,1 +1,0 @@
-from .mode import PredictionMode
