--- src/python/mini_face/gaze/__init__.py
@@ -1,1 +1,0 @@
-from ._gaze import Extractor, Result
