+++ vcpkg.json
@@ -1,6 +1,6 @@
 {
   "name": "mini-face",
-  "version-semver": "0.0.4",
+  "version-semver": "0.0.9",
   "dependencies": ["openblas", "opencv", "dlib", "pybind11"],
   "description": "A minimalist Python wrapper for OpenFace library"
 }
